+++ src/util/IconRow.java
@@ -8,12 +8,12 @@ public class IconRow extends InformationContainer{
 	
 	private static final int ICON_SIZE = 32;
 	private int amntIcons;
-	private byte[] icons;
+	private int[] icons;
 	private Event[] onClickEvents;
 	private Event[] events;
 	private int hMargin,vMargin;
 	
-	public IconRow(Event[] events, byte[] icons, int xSpan, int ySpan) {
+	public IconRow(Event[] events, int[] icons, int xSpan, int ySpan) {
 		
 		this.amntIcons = icons.length;
 		this.icons = icons;
@@ -30,7 +30,7 @@ public class IconRow extends InformationContainer{
 		int xOffset=hMargin;
 		int yOffset=vMargin;
 		int eventCtr = 0;
-		for(byte b : icons) {
+		for(int b : icons) {
 			int[] iconPixels = Resources.PORTRAITSx32.get(b);
 			//print(iconPixels,32,32);
 			fillWithGraphics(xOffset, xOffset+ICON_SIZE-1, yOffset, yOffset+ICON_SIZE-1, iconPixels, true);
